QLNTF-84 suggestion service delete service

+++ src/pages/service/index.tsx
@@ -1,5 +1,5 @@
 import { DeleteOutlined, EditOutlined, PlusOutlined } from '@ant-design/icons';
-import { Button, Checkbox, message, PageHeader, Space, Tooltip } from 'antd';
+import { Button, message, Modal, PageHeader, Space, Tooltip } from 'antd';
 import { ColumnsType } from 'antd/lib/table';
 import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
@@ -20,22 +20,27 @@ const ServicePage = () => {
         setSelectedRowKeys(newSelectedRowKeys);
     };
     const handleRemove = (_id?: string[]) => {
-        if (_id && _id?.length > 0) {
-            dispatch(deleteService({ data: _id }));
-        } else {
-            dispatch(deleteService({ data: selectedRowKeys }));
-        }
-        message.success(MESSAGES.DEL_SUCCESS);
+        Modal.confirm({
+            centered: true,
+            title: `Bạn có muốn xóa dịch vụ không!`,
+            cancelText: 'Hủy',
+            okText: 'Xóa',
+            onOk: async () => {
+                if (_id && _id?.length > 0) {
+                    dispatch(deleteService({ data: _id }));
+                } else {
+                    dispatch(deleteService({ data: selectedRowKeys }));
+                }
+                message.success(MESSAGES.DEL_SUCCESS);
+            },
+        });
     };
     const columnsService: ColumnsType = [
         {
             title: 'Tên dịch vụ',
             dataIndex: 'serviceName',
         },
-        {
-            title: 'Loại dịch vụ',
-            dataIndex: 'serviceTypeName',
-        },
+
         {
             title: 'Đơn giá',
             dataIndex: 'unitPrice',
@@ -43,17 +48,7 @@ const ServicePage = () => {
                 return <>{generatePriceToVND(price)}</>;
             },
         },
-        {
-            title: 'Đang dùng',
-            dataIndex: 'isActive',
-            render: (isActive) => {
-                return (
-                    <>
-                        <Checkbox checked={isActive} disabled />
-                    </>
-                );
-            },
-        },
+
         {
             title: '',
             width: '5%',
